pyproject.toml
[project]
name = "felix-exchange-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "websockets>=15.0.1",
    "asyncpg>=0.29.0",
    "aiohttp>=3.9.0",
    "msgpack>=1.1.2",
]
